Add basic OpenMPI container configuration
For having OpenMPI work, we shall let all containers know each other and
let them communicate through SSH. The current setup creates an
environment with all needed configuration to establish a SSH connection
by simply issuing

ssh <machine-name>

This is the very requirement for having the OpenMPI configured with SSH.
The same network configuration can be used for different algorithms
through this generic `node` image. Each algorithm must contain a `main`
executable which is buildable by `make`ing its directory once the
image is built.

A makefile in the root is provided to easy the execution of these
comands, including the image build through the `build` target, container
creation and configuration through the `setup` target, and algorithm
execution through `run` target. Note that this last command executes the
algorithm specified in the ALGORITHM environment variable, which
defaults to `hello`.