Use an explict queue to manage resource
When a node requests access to the resource, it is now put into a queue.
Whenever the resource is free, a node is poped and granted access to the
resource. To make this possible, we now implement async call to probing
the MPI message queues and treating them whenver a message exists.